+++ cli/src/lib.rs
@@ -1,3 +1,5 @@
 pub mod pubsub;
 pub mod stream;
+pub mod stream_grpc;
+pub mod stream_richat;
 pub mod track;